+++ src/lib/components/Modal/Modal.svelte
@@ -1,15 +1,21 @@
 <script lang="ts">
+	import ModalFooter from './ModalFooter.svelte';
+	import type { Footer } from './modal-types.js';
     import {clickOutside} from "../../directives/clickOutside.js";
     import { IconX } from '@hyvor/icons';
     import IconButton from './../IconButton/IconButton.svelte';
-	import { scale } from 'svelte/transition';
+	import { fade, scale } from 'svelte/transition';
     import { onMount, tick } from "svelte";
+    import Loader from "../Loader/Loader.svelte";
 
     export let show = false;
     export let title = "";
     export let size : "small" | "medium" | "large" = "medium";
     export let closeOnOutsideClick = true;
     export let closeOnEscape = true;
+    export let loading : boolean | string = false;
+
+    export let footer : null | Footer = null;
 
     let wrapEl: HTMLDivElement;
     let innerEl: HTMLDivElement;
@@ -30,7 +36,7 @@
 </script>
 
 <svelte:window on:keyup={e => {
-    if (e.key === 'Escape' && closeOnEscape) {
+    if (e.key === 'Escape' && closeOnEscape && !loading) {
         show = false;
     }
 }} />
@@ -40,15 +46,18 @@
     <div 
         class="wrap"
         bind:this={wrapEl}
+        in:fade={{duration: 100}}
+        out:fade={{duration: 100}}
     >
 
-        <div 
+        <div
             class="inner {size}"
             use:clickOutside={{
                 enabled: closeOnOutsideClick,
-                callback: () => show = false
+                callback: () => !loading ? show = false : null
             }}
             in:scale={{duration: 100, start: 0.9, opacity: 0.9}}
+            out:scale={{duration: 100, start: 0.9, opacity: 0.9}}
             bind:this={innerEl}
         >
 
@@ -77,9 +86,31 @@
                 <slot /> 
             </div>
 
-            {#if $$slots.footer}
+            {#if $$slots.footer || footer}
                 <div class="footer">
-                    <slot name="footer" />
+                    {#if $$slots.footer}
+                        <slot name="footer" />
+                    {:else if footer}
+                        <ModalFooter 
+                            {footer}
+                            bind:show={show}
+                            on:cancel
+                            on:confirm
+                        />
+                    {/if}   
+                </div>
+            {/if}
+
+            {#if loading}
+                <div 
+                    class="loading"
+                    in:fade={{duration: 100}}
+                >
+                    <Loader full>
+                        {#if typeof loading === "string"}
+                            {loading}
+                        {/if}
+                    </Loader>
                 </div>
             {/if}
 
@@ -89,6 +120,18 @@
 {/if}
 
 <style>
+
+    .loading {
+        position: absolute;
+        top: 0;
+        left: 0;
+        z-index: 1;
+        width: 100%;
+        height: 100%;
+        background-color: var(--box-background);
+        border-radius: var(--box-radius);
+        box-shadow: var(--box-shadow);
+    }
 
     .wrap {
         position: fixed;